REF: 関数移動、行削除など
