+++ command/Makefile
@@ -6,7 +6,7 @@
 #    By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2021/12/02 13:21:26 by mhirabay          #+#    #+#              #
-#    Updated: 2022/03/15 16:33:03 by mhirabay         ###   ########.fr        #
+#    Updated: 2022/03/15 20:32:29 by mhirabay         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -29,6 +29,7 @@ SRCNAME	:=	execute_builtin.c \
 			cmd_utils.c \
 			execve_error.c \
 			create_cmd_utils.c \
+			redirect_utils.c \
 
 SRCS	:= $(addprefix $(SRCDIR), $(SRCNAME))
