+++ command/src/exec_attribute_utils.c
@@ -3,13 +3,12 @@
 /*                                                        :::      ::::::::   */
 /*   exec_attribute_utils.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhirabay  <mhirabay@student.42tokyo.j      +#+  +:+       +#+        */
+/*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 14:50:54 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/03/15 14:50:54 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/03/15 20:23:57 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "command.h"
 
@@ -18,7 +17,6 @@ t_cmd	*get_cmd(t_exec_attr *ea)
 	return ((t_cmd *)ea->cmd_lst->content);
 }
 
-// コマンド名をgetする
 char	*get_cmd_name(t_list *cmd_lst)
 {
 	t_cmd	*cmd;
@@ -36,7 +34,6 @@ char	*get_argv_one(t_cmd *cmd)
 	}
 	if (cmd->args->next == NULL)
 	{
-		// ft_putstr_fd("no args", STDERR_FILENO);
 		return (NULL);
 	}
 	return (cmd->args->next->content);
