+++ command/src/execute.c
@@ -3,22 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   execute.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhirabay  <mhirabay@student.42tokyo.j      +#+  +:+       +#+        */
+/*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 14:50:54 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/03/15 14:50:54 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/03/15 20:33:30 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "command.h"
 
 int	execute_cmd(t_exec_attr *ea)
 {
 	int	pipe_count;
-	// void *leak;
 
-	// leak = malloc(1);
 	if (ea->cmd_lst == NULL)
 		return (0);
 	pipe_count = ft_lstsize(ea->cmd_lst) - 1;
@@ -32,19 +29,3 @@ int	execute_cmd(t_exec_attr *ea)
 	}
 	return (0);
 }
-
-// int	execute_cmd(t_list *cmd)
-// {
-// 	t_exec_attr	*e_attr;
-
-// 	(void)cmd;
-// 	int argc = 0;
-// 	const char **argv = NULL;
-// 	init(&e_attr);
-// 	create_cmd_from_arg(argc, argv, e_attr);
-// 	if (is_self_cmd(argv[1]))
-// 		execute_self(e_attr);
-// 	else
-// 		execute_builtin(e_attr);
-// 	return (0);
-// }
