+++ command/src/execute_self.c
@@ -3,13 +3,12 @@
 /*                                                        :::      ::::::::   */
 /*   execute_self.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhirabay  <mhirabay@student.42tokyo.j      +#+  +:+       +#+        */
+/*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 14:50:54 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/03/15 14:50:54 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/03/15 20:24:16 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "../includes/command.h"
 
@@ -23,8 +22,6 @@ bool	is_(char *command, char *cmd)
 
 void	execute_self_cmd(t_cmd	*c, t_exec_attr *ea, bool is_pipe)
 {
-	// TODO:argsだけ渡して成り立つものはargsの渡しで片付ける
-	// echo hoge fugaみたいなケースはargsの線形リストに入るので、線形リストを回さないとだめ
 	if (is_(CD, c->cmd))
 		g_exit_status = exec_self_cd(c, ea);
 	else if (is_(EXPORT, c->cmd))
@@ -43,7 +40,6 @@ void	execute_self_cmd(t_cmd	*c, t_exec_attr *ea, bool is_pipe)
 		exit(g_exit_status);
 }
 
-// TODO: is_関数を使う
 bool	is_self_cmd(char *cmd)
 {
 	if (is_(CD, cmd))
