+++ command/src/execve_error.c
@@ -3,13 +3,12 @@
 /*                                                        :::      ::::::::   */
 /*   execve_error.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhirabay  <mhirabay@student.42tokyo.j      +#+  +:+       +#+        */
+/*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 14:50:54 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/03/15 14:50:54 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/03/15 20:28:48 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "command.h"
 
@@ -19,10 +18,8 @@ static bool	can_exit(char *cmd_path)
 
 	if (stat(cmd_path, &stat_buf) == -1)
 		exit(EXIT_FAILURE);
-	// 所有者の実行許可を確認している
 	if ((stat_buf.st_mode & S_IXUSR) != S_IXUSR)
 		return (false);
-	// 所有者の読み込み許可を確認している
 	if ((stat_buf.st_mode & S_IRUSR) != S_IRUSR)
 		return (false);
 	return (true);
@@ -35,11 +32,10 @@ void	execve_error(int cp_errno, char *cmd_path)
 	if (cp_errno == ENOENT)
 		exit_status = 127;
 	else
-		exit_status = 126; // command not found以外は終了ステータスが126
-	// errnoがENOEXECでも実行権限がなければEACCESに変更する
+		exit_status = 126;
 	if (cp_errno == ENOEXEC && !can_exit(cmd_path))
 		cp_errno = EACCES;
-	else if (cp_errno == ENOEXEC) // ENOEXECは終了ステータス0
+	else if (cp_errno == ENOEXEC)
 		exit(EXIT_SUCCESS);
 	ft_put_cmd_error(cmd_path, strerror(cp_errno));
 	exit(exit_status);
