+++ command/src/free_pipe_attr.c
@@ -3,13 +3,12 @@
 /*                                                        :::      ::::::::   */
 /*   free_pipe_attr.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mhirabay  <mhirabay@student.42tokyo.j      +#+  +:+       +#+        */
+/*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 14:50:54 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/03/15 14:50:54 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/03/15 20:26:38 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #include "command.h"
 
@@ -28,7 +27,6 @@ static void	free_pipe_fd(t_pipe_attr *pa)
 
 void	free_pipe_attr(t_pipe_attr *pa)
 {
-	// ここでpa->current_cmdもfreeすべきかも
 	free_pipe_fd(pa);
 	free(pa->cpid_array);
 }
